+++ internal/app/forklift/pallets.go
@@ -202,21 +202,19 @@ func (p *FSPallet) LoadFSRepoReqs(searchPattern string) ([]*FSRepoReq, error) {
 
 // LoadPkgReq loads the PkgReq from the pallet for the package with the specified package path.
 func (p *FSPallet) LoadPkgReq(pkgPath string) (r PkgReq, err error) {
-	reposFS, err := p.GetRepoReqsFS()
-	if err != nil {
-		return PkgReq{}, errors.Wrap(err, "couldn't open directory for repo requirements from pallet")
-	}
 	if path.IsAbs(pkgPath) { // special case: package should be provided by the pallet itself
 		return PkgReq{
 			PkgSubdir: strings.TrimLeft(pkgPath, "/"),
 			Repo: RepoReq{
-				GitRepoReq{
-					RequiredPath: p.Def.Pallet.Path,
-				},
+				GitRepoReq{RequiredPath: p.Def.Pallet.Path},
 			},
 		}, nil
 	}
 
+	reposFS, err := p.GetRepoReqsFS()
+	if err != nil {
+		return PkgReq{}, errors.Wrap(err, "couldn't open directory for repo requirements from pallet")
+	}
 	fsRepoReq, err := loadFSRepoReqContaining(reposFS, pkgPath)
 	if err != nil {
 		return PkgReq{}, errors.Wrapf(err, "couldn't find repo providing package %s in pallet", pkgPath)